feat:delete user and create to archive table

+++ controllers/users.controller.ts
@@ -142,47 +142,46 @@ export async function deleteUser(req: Request, res: Response, next: NextFunction
       });
     }
 
-    if (existingUser.image) {
-      await deleteImageFromCloudinary(existingUser.image)
-    }
-
     // Archive and Delete Attendance
     if (existingUser.attendance && existingUser.attendance.length > 0) {
-      await Archive.createMany(existingUser.attendance.map((a: any) => ({
-        original_id: a.id,
-        table_name: "Attendance",
-        data: a,
-        deleted_by: "System",
-      })))
+      const attendArchive = await Archive.archiveAttendances(existingUser.attendance, req.user.id)
+      if (!attendArchive) {
+        return res.status(500).json({
+          success: false,
+          message: "Gagal mengarsipkan absensi.",
+        })
+      }
       await Attendance.deleteByUserId(id)
     }
 
     // Archive and Delete AttendanceRecap
     const recaps = await AttendanceRecap.getByUserId(id)
     if (recaps && recaps.length > 0) {
-      await Archive.createMany(recaps.map((r: any) => ({
-        original_id: r.id,
-        table_name: "AttendanceRecap",
-        data: r,
-        deleted_by: "System",
-      })))
+      const recapArchive = await Archive.archiveAttendanceRecaps(recaps, req.user.id)
+      if (!recapArchive) {
+        return res.status(500).json({
+          success: false,
+          message: "Gagal mengarsipkan recap absensi.",
+        })
+      }
       await AttendanceRecap.deleteByUserId(id)
     }
 
     // Archive User
     // eslint-disable-next-line @typescript-eslint/no-unused-vars
     const { attendance, ...userData } = existingUser
-    await Archive.create({
-      original_id: userData.id,
-      table_name: "Users",
-      data: userData as any,
-      deleted_by: "System"
-    })
+    const userArchive = await Archive.archiveUser(userData, req.user.id)
+    if (!userArchive) {
+      return res.status(500).json({
+        success: false,
+        message: "Gagal mengarsipkan user.",
+      })
+    }
 
-    const data = await User.deleteUser(id)
+    await User.deleteUser(id)
     res.status(200).json({
       success: true,
-      data
+      message: "User berhasil dihapus.",
     })
   } catch (error) {
     next(error)
